feat(invoices): default review reminder set when a quote becomes recurring
- The "to review" reminder is now shown only for quotes that have a
  subscription article and are in the "recurring" state.
- A backend trigger sets a default reminder when a quote becomes recurring:
  check once a year, at the start of the last month of the recurrence
  (derived from subscription_ends_at, falling back to the start month).
- Move the review-occurrence computation to the shared package so the trigger
  and the frontend component compute next_review_date the same way.
