+++ shared/src/invoices.ts
@@ -1,7 +1,76 @@
 import _ from "lodash";
 import { DateTime } from "luxon";
 import { getVatCode, standardCodeToVatValue } from "./consts";
-import { InvoicesBase as Invoices, InvoiceTotal } from "./types";
+import { InvoiceReview, InvoicesBase as Invoices, InvoiceTotal } from "./types";
+
+// ---------------------------------------------------------------------------
+// "To review" reminders (quotes): recurring rules made of a day-of-month spec
+// and a month spec. Shared by the frontend (component) and the backend (the
+// trigger that sets a default reminder when a quote becomes recurring).
+// ---------------------------------------------------------------------------
+
+// Resolve a reminder day-of-month spec to a timestamp (noon, to avoid TZ edges)
+const resolveReviewDay = (year: number, month0: number, day: string): number => {
+  const daysInMonth = new Date(year, month0 + 1, 0).getDate();
+  let d: number;
+  if (day === "first") d = 1;
+  else if (day === "last") d = daysInMonth;
+  else if (day === "middle") d = 15;
+  else d = Math.min(Math.max(parseInt(day) || 1, 1), daysInMonth);
+  return new Date(year, month0, d, 12, 0, 0, 0).getTime();
+};
+
+// All review occurrences (from every reminder rule) within [startTs, endTs]
+export const getReviewOccurrences = (
+  review: Partial<InvoiceReview> | undefined,
+  startTs: number,
+  endTs: number,
+): number[] => {
+  if (!review?.reminders?.length) return [];
+  const occurrences: number[] = [];
+  const startYear = new Date(startTs).getFullYear();
+  const endYear = new Date(endTs).getFullYear();
+  for (const reminder of review.reminders) {
+    if (!reminder?.day || !reminder?.month) continue;
+    for (let year = startYear; year <= endYear; year++) {
+      if (reminder.month === "every") {
+        for (let m = 0; m < 12; m++) {
+          occurrences.push(resolveReviewDay(year, m, reminder.day));
+        }
+      } else {
+        const month0 = (parseInt(reminder.month) || 1) - 1;
+        occurrences.push(resolveReviewDay(year, month0, reminder.day));
+      }
+    }
+  }
+  return _.uniq(occurrences)
+    .filter((t) => t >= startTs && t <= endTs)
+    .sort((a, b) => a - b);
+};
+
+const FIVE_YEARS_MS = 1000 * 60 * 60 * 24 * 366 * 5;
+
+// Next review date strictly after the given timestamp (null if none)
+export const getNextReviewDate = (
+  review: Partial<InvoiceReview> | undefined,
+  fromTs: number,
+): number | null => {
+  const occurrences = getReviewOccurrences(review, fromTs, fromTs + FIVE_YEARS_MS);
+  return occurrences.find((t) => t > fromTs) ?? null;
+};
+
+// Last review date strictly before the given timestamp (null if none)
+export const getPrevReviewDate = (
+  review: Partial<InvoiceReview> | undefined,
+  beforeTs: number,
+): number | null => {
+  const occurrences = getReviewOccurrences(
+    review,
+    beforeTs - FIVE_YEARS_MS,
+    beforeTs,
+  ).filter((t) => t < beforeTs);
+  return occurrences.length ? occurrences[occurrences.length - 1] : null;
+};
 
 export const applyOffset = (
   date: Date,
